getMentionedUser -> firstMentionOrFail
