+++ src/commands/addbye.ts
@@ -1,6 +1,6 @@
 import { CommandDefinition } from "../Command";
+import { firstMentionOrFail, reply } from "../util/discord";
 import { getLogger } from "../util/logger";
-import { reply } from "../util/reply";
 
 const logger = getLogger("command:addbye");
 
@@ -10,7 +10,7 @@ const command: CommandDefinition = {
 	executor: async (msg, args, support) => {
 		const [id] = args;
 		await support.tournamentManager.authenticateHost(id, msg.author.id);
-		const player = support.discord.getMentionedUser(msg);
+		const player = firstMentionOrFail(msg);
 		logger.verbose(
 			JSON.stringify({
 				channel: msg.channel.id,
