+++ src/commands/addhost.ts
@@ -1,6 +1,6 @@
 import { CommandDefinition } from "../Command";
+import { firstMentionOrFail, reply } from "../util/discord";
 import { getLogger } from "../util/logger";
-import { reply } from "../util/reply";
 
 const logger = getLogger("command:addhost");
 
@@ -10,7 +10,7 @@ const command: CommandDefinition = {
 	executor: async (msg, args, support) => {
 		const [id] = args;
 		await support.tournamentManager.authenticateHost(id, msg.author.id);
-		const newHost = support.discord.getMentionedUser(msg);
+		const newHost = firstMentionOrFail(msg);
 		logger.verbose(
 			JSON.stringify({
 				channel: msg.channel.id,
