--- src/util/reply.ts
@@ -1,8 +1,0 @@
-import { Message, Textable } from "eris";
-
-export async function reply(
-	msg: Message,
-	...args: Parameters<Textable["createMessage"]>
-): ReturnType<Textable["createMessage"]> {
-	return await msg.channel.createMessage(...args);
-}
